Initialize book class
