+++ models/label.rb
@@ -2,7 +2,7 @@ class Label
   attr_reader :id, :title, :color, :items
 
   def initialize(title, color, id = nil)
-    @id = id || Random.rand(1...10_000)
+    @id = id || Random.rand(1..10_000)
     @title = title
     @color = color
     @items = []
